Alerts Visable in the same PHP Page

+++ myblogSearch.php
@@ -115,42 +115,44 @@ require_once 'access.php';
             $encryptedValue = base64_decode($encodedCookie);
             $userId = openssl_decrypt($encryptedValue, 'aes-256-cbc', $encryptKey, 0, $encryptIV);
             
-            $sql = "SELECT blog_id, title, blog, filename FROM blog where user_id = '$userId'";
-            $result = $conn->query($sql);
-            if ($result->num_rows > 0) {  
-               // echo "<ul>";
-                
-                while ($row = $result->fetch_assoc()) 
-                {
-                    echo "<div class='mb-2 p-2 bg-light text-dark'>";
-                    echo "Title: <input class='form-control' type='text' value='" . $row["title"] . "' aria-label='readonly input example' readonly><br>";    
-                    echo "Blog: <textarea class='form-control' style='height: 150px' readonly>" . $row["blog"] . "</textarea><br>";
-                    echo "<img src='img/" . $row["filename"] . "' class='d-block w-50 h-50' alt='blog image'><br>";
-                    echo "<form action='delete.php' method='POST'>";
-                    echo "<input class='form-control' type='hidden' value='" . $row["blog_id"] . "' aria-label='readonly input example' readonly name='blog_id'>";
-                    echo "<button type='button' class='btn btn-danger' data-bs-toggle='modal' data-bs-target='#staticBackdrop'>Delete </button>";
-                    echo  " <div class='modal fade' id='staticBackdrop' data-bs-backdrop='static' data-bs-keyboard='false' tabindex='-1' aria-labelledby='staticBackdropLabel' aria-hidden='true'>
-                            <div class='modal-dialog'>
-                              <div class='modal-content'>
-                                <div class='modal-header'>
-                                  <h1 class='modal-title fs-5' id='staticBackdropLabel'>Confirm Delete</h1>
-                                  <button type='button' class='btn-close' data-bs-dismiss='modal' aria-label='Close'></button>
-                                </div>
-                                <div class='modal-body'>
-                                  Are you sure you want to Delete this Blog?
-                                </div>
-                                <div class='modal-footer'>
-                                  <button type='button' class='btn btn-secondary' data-bs-dismiss='modal'>No</button>
-                                  <button type='submit' class='btn btn-danger'>Yes, Delete</button>
+            if(isset($_POST['searchTerm'])){
+              $searchTerm = $_POST['searchTerm'];
+              $sql = "SELECT blog_id, title, blog, filename FROM blog where user_id = '$userId' AND title LIKE '%$searchTerm%'";
+              $result = $conn->query($sql);
+              if ($result->num_rows > 0) {  
+                // echo "<ul>";
+                  
+                  while ($row = $result->fetch_assoc()) 
+                  {
+                      echo "<div class='mb-2 p-2 bg-light text-dark'>";
+                      echo "Title: <input class='form-control' type='text' value='" . $row["title"] . "' aria-label='readonly input example' readonly><br>";    
+                      echo "Blog: <textarea class='form-control' style='height: 150px' readonly>" . $row["blog"] . "</textarea><br>";
+                      echo "<img src='img/" . $row["filename"] . "' class='d-block w-50 h-50' alt='blog image'><br>";
+                      echo "<form action='delete.php' method='POST'>";
+                      echo "<input class='form-control' type='hidden' value='" . $row["blog_id"] . "' aria-label='readonly input example' readonly name='blog_id'>";
+                      echo "<button type='button' class='btn btn-danger' data-bs-toggle='modal' data-bs-target='#staticBackdrop'>Delete </button>";
+                      echo  " <div class='modal fade' id='staticBackdrop' data-bs-backdrop='static' data-bs-keyboard='false' tabindex='-1' aria-labelledby='staticBackdropLabel' aria-hidden='true'>
+                              <div class='modal-dialog'>
+                                <div class='modal-content'>
+                                  <div class='modal-header'>
+                                    <h1 class='modal-title fs-5' id='staticBackdropLabel'>Confirm Delete</h1>
+                                    <button type='button' class='btn-close' data-bs-dismiss='modal' aria-label='Close'></button>
+                                  </div>
+                                  <div class='modal-body'>
+                                    Are you sure you want to Delete this Blog?
+                                  </div>
+                                  <div class='modal-footer'>
+                                    <button type='button' class='btn btn-secondary' data-bs-dismiss='modal'>No</button>
+                                    <button type='submit' class='btn btn-danger'>Yes, Delete</button>
+                                  </div>
                                 </div>
                               </div>
-                            </div>
-                          </div>";   
-                    echo "</form>";
-                    echo "</div>";
-                    
+                            </div>";   
+                      echo "</form>";
+                      echo "</div>";
+                      
+                    }
                   }
-                
                // echo "</ul>";
             } else {
                 echo "No blogs found.";
